% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundancea_and_occupancy_core.R
\name{abundance_and_occupancy_core}
\alias{abundance_and_occupancy_core}
\title{Proportion of Reads and Replicates:}
\usage{
abundance_and_occupancy_core(
  otu_table,
  prop_rep = 0.5,
  prop_reads = 2e-04,
  taxa_as_rows = TRUE
)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where the first row is the OTU ID and column names refer to sites}

\item{prop_reads}{percentage of reads}

\item{taxa_as_rows}{\code{logical} data must be in a format where the taxa are rows
and the sites are columns. The default value is \code{TRUE},
if \code{FALSE} data will be transposed for downstream analysis.}

\item{occupancy_core}{percentage of replicates}
}
\value{
the names of OTUs which meet the proportion of reads and replicate criteria
}
\description{
This method assigns taxa to the core
if they account for some proportion of the total reads for the sequencing run
and if they are present in at least x% of the total number of replicates.

In this example, a core taxa must account for 0.02% of the total reads
for the entire otu table and be present in at least 50% of sites.
}
\examples{
abundance_and_occupancy_core(arabidopsis)

}
