% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arabidopsis_m21_4.R
\docType{data}
\name{arabidopsis_m21_4}
\alias{arabidopsis_m21_4}
\title{Descritpion for the Arabidopsis phyloseq objects}
\format{
An object of class \code{phyloseq} of length 1.
}
\source{
\href{https://www.nature.com/articles/nature11237}{Nature}
}
\usage{
data(arabidopsis_m21_4)
}
\description{
Data derived from Lundberg et al. 2012. Dataset used for core assignment comparisons is a subset of the entire
core Arabidopsis thaliana root microbiome dataset. We utilized only the rhizosphere (r) samples from the m21 site
for our comparisons of core assignment methods. Methods for filtering of the original dataset to include only
M21 rhizosphere (M21_R) samples can be found in data-raw/arabidopsis.R. The entire raw arabidopsis datset is
named arabidopsis_ps. The M21_R subset of the arabidopsis dateset is named arabidopsis_m21_r_ps.
}
\examples{
data(arabidopsis_m21_4)
}
\references{
D.S. Lundberg, S.L. Lebeis, S.H. Paredes, S. Yourstone, J. Gehring, S. Malfatti, J. Tremblay, A. Engelbrektson,
V. Kunin, T.G. del Rio, R.C. Edgar, T. Eickhorst, R.E. Ley, P. Hugenholtz, S.G. Tringe, J.L. Dangl, Defining the core
Arabidopsis thaliana root microbiome, Nature. 488 (2012) 86–90. doi:10.1038/nature11237.
(\href{https://www.nature.com/articles/nature11237}{Nature})

What does QTL Archive thing mean?
}
\keyword{datasets}
