% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_core.R
\name{common_core}
\alias{common_core}
\title{Using the core_methods output,
return a dataframe of taxa IDs,
the number of core methods they are included in,
and the name of the core methods of inclusion}
\usage{
common_core(data, threshhold = 0)
}
\arguments{
\item{data}{the \code{core_methods} output of your taxa table}

\item{threshhold}{number of methods to include in results, all methods included by default}
}
\description{
Using the core_methods output,
return a dataframe of taxa IDs,
the number of core methods they are included in,
and the name of the core methods of inclusion
}
\examples{
\dontrun{
core_methods(arabidopsis) \%>\% common_core(threshhold = 2)
}
}
