% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_methods.R
\name{core_methods}
\alias{core_methods}
\title{Combine Methods}
\usage{
core_methods(otu_table, taxa_as_rows = TRUE)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where
the first row is the OTU ID and column names refer to sites}

\item{taxa_as_rows}{\code{logical} data must be in a format where the taxa are rows
and the sites are columns. The default value is \code{TRUE},
if \code{FALSE} data will be transposed for downstream analysis.}
}
\value{
a dataframe with all observed taxa,
their inclusion to the core by method (delineated as a 1 or 0),
the mean, variance, and coefficient of variation.
}
\description{
A wrapper for
\code{summarise_taxa}
\code{prop_reads},
\code{occupancy_core},
\code{abundance_and_occupancy_core} and
\code{hard_cutoff} methods
}
\examples{
\dontrun{core_methods(arabidopsis)}

}
