% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_venn.R
\name{core_venn}
\alias{core_venn}
\alias{core_venn.default}
\alias{core_venn.core_methods}
\title{Core Venn Diagram}
\usage{
core_venn(
  combined_taxa_data,
  category_names = c("Proportion of \\n Sequence Reads",
    "Proportion of \\n Sequence Replicates",
    "Proportion of \\n Sequence Reads and Replicates", "Hard Cut Off"),
  low = "white",
  high = "#7E191B"
)

\method{core_venn}{default}(
  combined_taxa_data,
  category_names = c("Proportion of \\n Sequence Reads",
    "Proportion of \\n Sequence Replicates",
    "Proportion of \\n Sequence Reads and Replicates", "Hard Cut Off"),
  low = "white",
  high = "#7E191B"
)

\method{core_venn}{core_methods}(
  combined_taxa_data,
  category_names = c("Proportion of \\n Sequence Reads",
    "Proportion of \\n Sequence Replicates",
    "Proportion of \\n Sequence Reads and Replicates", "Hard Cut Off"),
  low = "white",
  high = "#7E191B"
)
}
\arguments{
\item{combined_taxa_data}{the taxa data of type \code{core_methods}}

\item{category_names}{the names of the core inclusion methods.}

\item{low}{the low hexedicmal color}

\item{high}{the high hexidecimal color}
}
\value{
a venn diagram of the taxa included within each core method
}
\description{
Create a venn diagram using the output of core_methods to find which core taxa are included
in which methods
}
\examples{
\dontrun{
core_arabidopsis <- core_methods(arabidopsis)
core_venn(core_arabidopsis)
}
}
