% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_cutoff.R
\name{hard_cutoff}
\alias{hard_cutoff}
\title{Hard Cut-Off}
\usage{
hard_cutoff(otu_table, cutoff = 25, sites = 5)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where the first row is the OTU ID and column names refer to sites}

\item{cutoff}{the threshhold minimum number of reads - defaults to 25}

\item{sites}{the theshhold minimum number of sites - defaults to 5}
}
\value{
the names of OTUs which meet the hard cut off criteria
}
\description{
This method assigns taxa to the core
if they are present in more than a pre-determined number of sites
and have a total abidance greater than a pre-determined number of reads.
In our example we set the minimum number of sites to 5 and the minimum number of reads to 25.
Here we use the hard cut off described in Lundberg (2012), but realize this is any threshold.
}
\examples{
prop_reads(arabidopsis)

}
