% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_ps-data.R
\docType{data}
\name{human_ps}
\alias{human_ps}
\title{Descritpion for human microbiome project phyloseq objects.}
\format{
An object of class \code{phyloseq} of length 1.
}
\source{
\href{https://hmpdacc.org/ihmp/}{https://hmpdacc.org/ihmp/}

\href{http://downloads.hmpdacc.org/data/HMQCP/otu_table_psn_v35.txt.gz}{http://downloads.hmpdacc.org/data/HMQCP/otu_table_psn_v35.txt.gz}

\href{http://downloads.hmpdacc.org/data/HMQCP/v35_map_uniquebyPSN.txt.bz2}{http://downloads.hmpdacc.org/data/HMQCP/v35_map_uniquebyPSN.txt.bz2}
}
\usage{
data(human_ps)
}
\description{
This datsaet was used for comparison of core assignment methods. This human_ps phyloseq object contains all samples
from the human microbiome project. Methods for filtering of the original dataset to include only stool
samples can be found in data-raw/human.R. The entire raw human datset is named human_ps. 
The stool subset of the human dateset is named human_working_set.
}
\examples{
\dontrun{data(human_ps)}
}
\references{
Consortium, Human Microbiome Project, and others. 2012. “Structure, Function and Diversity of the Healthy Human Microbiome.” 
Nature 486 (7402). Nature Research: 207–14.
}
\keyword{datasets,}
\keyword{human}
\keyword{microbiome,}
\keyword{stool}
