% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-data.R
\docType{data}
\name{human_stool}
\alias{human_stool}
\title{Descritpion for subset of the human microbiome dataset used for comparison of core assignment methods. 
Methods for filtering of the original dataset can be found in data-raw/human.R.}
\format{
An object of class \code{data.frame} with 11752 rows and 319 columns.
}
\source{
\href{https://hmpdacc.org/ihmp/}
}
\usage{
data(human_stool)
}
\description{
Descritpion for subset of the human microbiome dataset used for comparison of core assignment methods. 
Methods for filtering of the original dataset can be found in data-raw/human.R.
}
\examples{
data(human_stool)
core_methods(human_stool)
}
\references{

}
\keyword{datasets}
