% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_reads.R
\name{prop_reads}
\alias{prop_reads}
\title{Proportion of Reads:}
\usage{
prop_reads(otu_table, perc = 0.75)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where the first row is the OTU ID and column names refer to sites}

\item{perc}{blah blah}
}
\value{
the names of OTUs which meet the proportion of reads criteria
}
\description{
This method assigns taxa to the core if they are in the top X% of reads.
Taxa are ranked in abundance and the cumulative sum is recorded.
Any taxa which appears before some cutoff percentage is included in the core.
In this example, a taxa will be assigned to the core if they account for the first 75% of the reads
}
\examples{
prop_reads(arabidopsis)

}
