% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_reads_and_reps.R
\name{prop_reads_and_reps}
\alias{prop_reads_and_reps}
\title{Proportion of Reads and Replicates:}
\usage{
prop_reads_and_reps(otu_table, sites = 0.5)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where the first row is the OTU ID and column names refer to sites}

\item{sites}{the percent of sites the OTU needs to be present in}
}
\value{
the names of OTUs which meet the proportion of reads and replicate criteria
}
\description{
This method assigns taxa to the core if they account for some proportion of the total reads for the sequencing run and if they are present in at least x% of the total number of replicates. In this example, a core taxa must account for 0.01% of the total reads for the entire otu table and be present in at least 50% of sites.
}
\examples{
prop_reads(arabidopsis)

}
