% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_reads.R
\name{sum_reads}
\alias{sum_reads}
\title{Summation of Reads:}
\usage{
sum_reads(otu_table, readn = 0.75, taxa_as_rows = TRUE)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where the first row is the OTU ID and column names refer to sites}

\item{readn}{count of reads needed}

\item{taxa_as_rows}{\code{logical} data must be in a format where the taxa are rows
and the sites are columns. The default value is \code{TRUE},
if \code{FALSE} data will be transposed for downstream analysis.}
}
\value{
the names of OTUs which meet the proportion of reads criteria
}
\description{
This method assigns taxa to the core if they are in the top X% of reads.
Taxa are ranked in abundance and the cumulative sum is recorded.
Any taxa which appears before some cutoff percentage is included in the core.
In this example, a taxa will be assigned to the core if they account for the first 75% of the reads
}
\examples{
sum_reads(arabidopsis)

}
