% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_OTUs.R
\name{summarize_OTUS}
\alias{summarize_OTUS}
\title{Calculate the mean, variance and CV of each OTU}
\usage{
summarize_OTUS(otu_table)
}
\arguments{
\item{otu_table}{a dataframe of OTUs
where the first row is the OTU ID and column names refer to sites}
}
\value{
a dataframe with the Mean, Variance and CV for each OTU
}
\description{
Calculate the mean, variance and CV of each OTU
}
\examples{
summarize_OTUS(arabidopsis)

}
