% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_taxa.R
\name{transpose_taxa}
\alias{transpose_taxa}
\title{If data are arranged where taxa are columns and sites rows
we need to transpose the data for downstream analysis}
\usage{
transpose_taxa(taxa_table)
}
\arguments{
\item{taxa_table}{the table to transpose}
}
\description{
If data are arranged where taxa are columns and sites rows
we need to transpose the data for downstream analysis
}
