% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_methods.R
\name{core_methods}
\alias{core_methods}
\title{Combine Methods}
\usage{
core_methods(otu_table)
}
\arguments{
\item{otu_table}{a dataframe of OTUs where
the first row is the OTU ID and column names refer to sites}
}
\value{
a dataframe with all observed taxa,
their inclusion to the core by method (delineated as a 1 or 0),
the mean, variance, and coefficient of variation.
}
\description{
A wrapper for
\code{summarise_taxa}
\code{prop_reads},
\code{prop_reps},
\code{prop_reads_and_reps} and
\code{hard_cutoff} methods
}
\examples{
core_methods(arabidopsis)

}
